feat: move concreteApp to concrete/main.go
